+++ src/headers/drawable.hpp
@@ -2,7 +2,7 @@
 #include <SDL.h>
 #include <SDL_image.h>
 #include "component.hpp"
-#include "vector2f.hpp"
+#include "vector2.hpp"
 
 //abstract class for each object which can be drawn
 class Drawable : public Component {
@@ -10,10 +10,10 @@ protected:
     SDL_Renderer* ren; // renderer for drawing
     SDL_Texture* text; // texture object of this class
     const char* pathToImage; // path to texture image
-    Vector2f textSize;
+    Vector2i textSize;
 public:
     // creates texture according to image path
-    bool init(GameObjectShar go_, Vector2f size_, SDL_Renderer* ren_, const char* pathToImage_);
+    bool init(GameObjectShar go_, Vector2i size_, SDL_Renderer* ren_, const char* pathToImage_);
     void draw(); // drawing update function
     void update() override {}
 };
