+++ src/headers/physic.hpp
@@ -1,6 +1,6 @@
 #pragma once 
 #include <SDL.h>
-#include "vector2f.hpp"
+#include "vector2.hpp"
 #include "game_object.hpp"
 #include <memory>
 #include <iostream>
@@ -10,7 +10,7 @@ typedef std::unique_ptr<SDL_Rect> RectPtr;
 class Physic : public Component {
 private:
 protected:
-    Vector2f size; // collider hitbox
+    Vector2i size; // collider hitbox
     Vector2f dir; // direction of force
     float force; // force value
     
@@ -31,11 +31,11 @@ public:
     //  1 4 3
     //    2
     // returns 0 if collision is upside, 1 if leftside, 2 if downside and 3 if rightside
-    static int isCollide(Vector2f fPos, Vector2f fSize, Vector2f sPos, Vector2f sSize);
+    static int isCollide(Vector2i fPos, Vector2i fSize, Vector2i sPos, Vector2i sSize);
     static int isCollide(SDL_Rect first, SDL_Rect second);
 
     bool init(GameObjectShar go_);
-    bool init(GameObjectShar go_, Vector2f size_, Vector2f dir_, float force_);
+    bool init(GameObjectShar go_, Vector2i size_, Vector2f dir_, float force_);
     void update () override;
     void print();
 };
